#38 Refactor TextDisplay and TextReader components; implement useTextDisplay hook for improved state management and readability; replace hardcoded text with textReaderTexts module for better localization support.
