+++ src/features/lireFlow/text/textReaderTexts.ts
@@ -1,0 +1,28 @@
+import { ROUTE_PATHS } from "../../../config/routes.ts";
+
+export const textReaderTexts = {
+uploadText: "Arraste o arquivo",
+uploadText2: ".txt",
+uploadText3: "até aqui ou ",
+uploadText4: "Selecione um arquivo",
+uploadText5: " ou ",
+placeholderSelectText: "Selecione um dos textos",
+texts: [
+"O Ratinho Rói-Rói (Leitura fácil)",
+"Conto ou não conto (Leitura média)",
+"A Cartomante (Leitura avançada)",
+],
+fileText: "Arquivo:",
+newText: "Escolher um novo texto",
+summary: {
+title: "LireFlow",
+texts: [
+"O objetivo desta tarefa é treinar a fluência leitora, com foco no ritmo e na automatização da leitura.",
+"É possível carregar um arquivo no formato .txt ou escolher um dos textos disponíveis na plataforma.",
+"Durante a tarefa, o texto será exibido com marcações com destaque palavra por palavra, em um tempo que pode ser ajustado conforme a necessidade.",
+],
+linkText: "Para mais informações, acesse o ",
+link: "Guia do Usuário.",
+linkRef: ROUTE_PATHS.USER_GUIDE_LIRE_FLOW,
+},
+};
